requirements.txt
django==2.0.5
bootstrap-admin==0.3.9
django-allauth==0.36.0